+++ src/lib/@types/animations.ts
@@ -1,0 +1,7 @@
+export interface RippleProps {
+	component?: string; // The component name to use for the ripple shape
+	center?: boolean;
+	color?: string; // CSS color
+	duration?: number; // In ms
+	disabled?: boolean; // Should the ripple be disabled
+}
