+++ src/lib/@types/core.ts
@@ -1,0 +1,3 @@
+export type LapikitPreprocessOptions = {
+	plugins?: string[];
+};
