+++ src/lib/@types/events.ts
@@ -1,0 +1,4 @@
+export type ClickOutsideOptions = {
+	exclude?: (HTMLElement | PointerEvent | null)[];
+	onClose: () => void;
+};
