+++ src/lib/@types/index.ts
@@ -1,0 +1,6 @@
+export * from './dom';
+export * from './components';
+export * from './props';
+export * from './events';
+export * from './animations';
+export * from './core';
