+++ src/lib/@types/props.ts
@@ -1,0 +1,3 @@
+export type PropValue = string | boolean | number | null | undefined;
+export type SizeType = 'default' | 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+export type RoundedType = 0 | '0' | 'xs' | 'sm' | 'md' | 'lg' | 'xl';
