--- src/lib/compiler/types/dom.ts
@@ -1,2 +1,0 @@
-export type SClassProp = string | string[] | Record<string, boolean | string> | undefined;
-export type SStyleProp = Record<string, boolean | string> | undefined;
