--- src/lib/compiler/types/index.ts
@@ -1,2 +1,0 @@
-export type * from './dom';
-export * from './options';
