+++ src/lib/components/appbar/appbar.types.ts
@@ -1,4 +1,4 @@
-import type { Component, RoundedType } from '$lib/utils/types';
+import type { Component, RoundedType } from '$lib/@types';
 
 type Density = 'compact' | 'comfortable' | 'default';
 type Variant = 'outline' | 'text';
