+++ src/lib/components/modal/modal.types.ts
@@ -1,4 +1,4 @@
-import type { Component, RoundedType } from '$lib/utils/types';
+import type { Component, RoundedType } from '$lib/@types';
 
 export type ModalSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'full';
 export type ModalPosition = 'top' | 'center' | 'bottom';
