+++ src/lib/components/spacer/spacer.types.ts
@@ -1,4 +1,4 @@
-import type { Base } from '$lib/utils/types';
+import type { Base } from '$lib/@types';
 
 export interface SpacerProps extends Base {
 	is?: 'div';
